feat: #10 1. UIView와 wave 의 동적 높이 맞추기 guage view - floating view 합치기     1. UIView와 wave 의 동적 높이 맞추기
