--- project-Emotion/project-Emotion/Controller/ViewController.swift
@@ -1,33 +1,0 @@
-//
-//  ViewController.swift
-//  project-Emotion
-//
-//  Created by Junhong Park on 2021/09/10.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    let gaugeView = GaugeWaveAnimationView(frame: UIScreen.main.bounds)
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-        view.addSubview(gaugeView)
-        gaugeView.startWaveAnimation()
-    }
-    
-    override func viewDidLayoutSubviews() {
-        super.viewDidLayoutSubviews()
-        gaugeView.setGradientLayer()
-    }
-    
-    override func viewDidDisappear(_ animated: Bool) {
-        super.viewDidDisappear(animated)
-        gaugeView.stopWaveAnimation()
-    }
-
-
-}
-
